requirements.txt
discord.py[voice]
discord-py-slash-command
lavalink
psutil
aiohttp
aiosqlite
aiomysql
requests